+++ src/views/index.ts
@@ -4,3 +4,4 @@ export { CommitHistoryManager } from './commitHistoryManager';
 export { VisualFileHistoryManager } from './visualFileHistoryManager';
 export { CommitGraphManager } from './commitGraphManager';
 export { LaunchpadManager } from './launchpadManager';
+export { RepoHealthManager } from './repoHealthManager';